requirements.txt
pipecat-ai[cartesia,openai,silero,deepgram]
fastapi
uvicorn
python-dotenv
loguru
plivo
python-multipart
jinja2